Add SignClient wrapper to manage WC sessions
The DAppConnector manages a list of signers to emit requested operations
via Wallet Connect. Additionally, it simplifies the process of
connecting, disconnecting, and restoring sessions by wrapping
SignClient.

DAppSigner is ready to send a request to the wallet for an operation.
Currently, it supports two operations - 'signMessages' and
'signTransaction'.

Added a new test to check the connection. It needs to be run with the
--verbose option to be able to interact with the test
(copy/paste the pairing string). The `test:connector` command for npm
has been added.